Update countrycitystate for local use
The original is intended to be used with node.js, which is not yet a
requirement for the project.

The purpose of the package is to populate the drop-down list for the
user to select a country and, when needed, a state or territory.

Convert the original API version into an object with methods, and add
new methods for creating HTML <option> elements.

Add a table for choosing the proper "prompt" for the state/territories
<option> based on the selected country.

Changed the "name" of 'Åland' to 'Aland' so that the name will sort in
the drop down where most "English" speakers would expect it. The single
'Å' at character position 51,933 was changed to an 'A'.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>