Convert the index file to php
Add the .htaccess file, the bulk of which is my standard template, with
the final pair of lines designed to make the system load the index.php
file for all requests, unless a specific file which _does_ exist is in
the URL.

The rendered page will be the sample blog page, until the point that I
need to use variable pages. It is long enough to fill most screens, and
has sections which can be jumped to on the left menu.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>