Move favorite icons to assets
Clean up the root directory by moving the favicons, and related images,
to a directory in assets. Update the rendered HTML to adjust.

Add a plain .ico file, somehow missed in the original copy of the site.

The favicon.ico file is kept in the root directory since that's where
everyone looks for it when linking to the site without a rendered page
to display. All the other icons can be moved since they are only linked
after viewing a rendered page and the client will have their full path.

Dropped the now defunct file index.html

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>