Setup for PHP processing
Add two files for normal use: include.php and conf.php.

The include.php is for configuring the site and its installation. File
paths owner and site information, and other data which may be different
between individual installed instances of the site. The conf.php file
is for configuring how the collected PHP files operate, and may be
different between sites while not required to be.

The conf.php file also sets the include path for the rest of the PHP
files, using the path defined in the include.php file. Part of that is
the inclusion of 'chindraba.php' which is for placing PHP code that
will not be needed in production while being needed during development.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>