Build new method for system init
The conf.php file has been renamed to definePaths.php and moved to the
script root from the web root. The only purpose I could see, going
forward, was for defining extra file paths internally, all of which are
based on the root of the script path, provided in the includes.php,
renamed from include.php.

Dropped the idea of using, and setting, the include path within PHP,
using instead constants for the path names with absolute paths instead.

Add index.html files to directories, which shouldn't be served by the
webserver anyway, to prevent any chance of a directory listing being
shown to client browsers.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>