Convert quote style
Using double quotes in PHP allows for using variables directly in the
creation of strings. Using single quotes within the HTML attributes has
the benefit of not having to escape the quotes most of the time.

In the process of converting the quote style, I also converted the ones
in the displayed content to HTML entities. This is mostly wasted effort
as the content will actually be rendered from a programatic source and
not from these files.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>