Setup for processing lists of strings in HTML
The plan is to generate the HTML in lists, allowing me to leverage the
ability of PHP to handle jagged arrays of arbitrary depth.

The functions will return lists, which may be added to a list without
being flattened. In the end, or at some other time, a list can be made
into a single level using the `flatten_list` function.

Prior to being sent to the client a list is converted into a single
string using the `render-list` function, which accepts any list to an
arbritary depth.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>