Extract page name from URL
The query string of the URL should give the page name. The `.htaccess`
file converts path names in the URL into a query string for any path
not found in the server's root silently, leaving the requested URL in
the browser address bar. Keeping the URL as the given path rather than
the converted query string is, hopefully, to keep search engines happy
and make SEO better.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>