Make supporting functions for building HTML nodes
Much of the non-value content will be within groups of elements. Making
the creation of source data manageable a data structure is going to be
used, the elements of which will be developed as the project grows. The
basis is an associative array, the base level identifying the type of
component, not the type of node, being built. At the base level will
also be a `common` element which applies to everything in the group and
reduces repetition. The collection of data might have some "content",
but the vast majority of the data should be meta data for the building
of the nodes.

Attributes known to be needed are the HTML attributes of `id`, `class`,
`type`, and the node type, or HTML 'tag' to build. Often the `id` will
be used in forming a compound `id` for built nodes and the `class` will
be a merged list of class names. There will also be, often, a default
value for some attributes, either based on where it's being used, or
what it is. During processing of the data of a given type of component,
the data in `common` will be used as the default, and be overridden by
the same attributes in the component's specific data.

So far, a working example, subject to change, is:
    $meta_data = [
        'common' => [
            'class-list' => 'class-one class-two',
            'id' => 'sample-item',
        ],
        'switch' => [
            'type' => 'checkbox',
            'checked' => false,

        ],
        'trigger' => [
            'class-list' => ['extra-class', 'another-class'],
        ]
    ];

Other attributes, and more intense processing will follow. The first to
be handled is the class list. Classes and roles are unusual in the HTML
attribute list as they can hold multiple values, in a space-separated
format. The processing of the role attribute can be handled in the same
fashion, perhaps even by the same function, later.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>