Add usage of new functions to check operations
Flesh out the favicons using the new functions to build the HTML rather
than pure text. It is longer but it checks some of the operations of
the building process.

Reorder the included files in functions-general.php to make it easier
to track what's been done for development. The list will need to be
split into different files, each reduced to a single collection of
routines. The order of the list now will make no difference in that
process then.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>