Add function to combine meta data for a node
The data for a node is a combination of what's declared for the node
within the group and the common data for the group. The functions which
build the node will have some attributes which are required, and which
have typical, or default values, and there may be something which has
the need to force an attribute's value. All these sources need to be
assembled into one assoc array of meta data for building the requested
node.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>