Rename temporary rendering files
The functions which are currently in individual files will be moved to
a single file, or small set of files, in preparation for the production
ready move. At this time having the file name match the function name
should make it simpler to track what's been done, and where to find the
information needed to make other functions work.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>