Document metadata structure and usage
In the process of writing the documentation I decided that the classes
and roles should be treated identically. Updated the related files to
make it so.

The notes on the metadata will be growing as more elements are added to
the processing system. At this point it is ready for creating switches.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>