Change names to match new targets
The system I have in mind to create will generate standardized values
for id and class names. Looking toward that, eventually, I've changed
the names of the elements connected to the switch building, as much as
I can predict based on what I hope to do, to reflect those names.

The CSS has been updated to reflect that as well. In time I hope to fix
the CSS to have fewer compound rules, excepting those connected to the
implementation of settings and menus, using more atomic style rules.
The entirity of atomic style CSS seems to be overboard, as is the BEM
system. Much of the concept, however, can be used where it seems to be
a good fit. To a small degree, BEM can also be used, in spirit. My use
is more a blend of atomic, BEM, and semantic, with intelligence applied
in keeping it useful, scalable, and workable.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>