Move data to fake database
The final form is yet to be built for the database. The access to it
cannot be built without the decisions around the database being made.
The hoped-for format of the data from the database can be set, as a
target if nothing else.

The fakeDB.php file simulates the return from the database of what is
hoped to be a good fit with room to grow. The call to the database will
be from the source-meta.php file. A similar setup for menu data, once I
get there, will be used, probably with the same fakeDB.php file.

The make_switch function was doing needless work, now removed.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>