Normailze nomeclature for metadata
Adjust the names used in processing the database data and in the docs
for the metadata so that the "words" used have, hopefully, consistent
context in the docs, the data processing and the use of that data.

Change the documentation from .php to .md. Better suited for such use.

The general break down, at this point is:
-  Group or UI Group: a collection of UI items, large or small, to
achieve a common purpose. A frequent group could be a collapsable menu,
with a 'switch' to track that status of the menu, a 'trigger' to open
and close it, and a 'menu' for holding the visual part of the menu. The
menu may be in a visual box, which may be in a container node and the
trigger could be included in the menu box, or have its own box. Each
facet of the set is covered by a UI item in the metadata. In addition,
as a general rule the data for the entire group is passed to, and from,
the functions to deal with any portion of it.
-  Item or UI Item: something which may be a single node. The UI item
'switch' is rendered as a single <input> in HTML. May also me a complex
section of the DOM tree. A 'menu' is rendered as a collection of links
in a list, which may be a section of another menu.
-  Element: one string-indexed element in an assoc array in the code.
This may also sometimes be used in reference to an HTML element
(a node), or some "visual" element on the rendered page, such as a
menu, footer, or other group of "visual" elements, which is probably
best refered to as an UI item. Hopefully both cases will be minimal.
-  Node: an HTML tag, such as <div> which may, or may not, contain more
nodes.
-  Attribute: Commonly an HTML attribute, often as a pair with the
value of same: such as `class="class-one other-class"` or
`type="checkbox"`. Sometime refering to an element of the lowest level
data in an assoc array, such as the checked attribute of a switch item.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>