Switch to locally available images
The two images presented in the value content were being loaded from
imgur.com. As continued off-line development progressed it got to be
frustrating waiting for the connection to time-out on page load. I had
the goto500 image handy, and added it to the site's files. The flow
diagram is somewhere in my system but not worth the effort to dig out
at the moment, so I merely commented out the use of it.

At this point what the value content _is_ does not matter, only that it
is there to fill the page. I ought to be on-line again when I get to a
point where the actual content of the pages is important.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>