Modify styles for burger menu icon operation
Add `widescreen.css` file for the CSS rules to make the widescreen
toggle work while all the `@media` rules in the regular CSS files which
deal with display width are commented out. The operation is that the
smallest version is presented when the toggle is off and the widest is
presented when the toggle is on. On narrow display settings with the
toggle on there are some weird effects, which can be discounted as they
are handled under normal use by the graduated `@media` queries used
when the override is not present. Lastly, as a clue that the switch is
in use the background of the logo is modified to show selected or not.

Add `colors.css` file for new color variable names and rules. For now
these will be hard-coded values, using the same structure as the scheme
files use. Once the color needs are established the scheme files can be
reworked to use the new names, and values if any.

Add a CSS file for new, or modified, styles. Included in this are the
styles as they are needed for changes to the produced HTML. Eventually
these changes _might_ be split into separate files. Until, and unless,
it is split the new styles are in `site-wide.css`.

Add all new CSS files to the rendered HTML.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>