Convert side menu containers
Change style names to be, in my opinion, more semantic.

The `flex-direction: row-reverse` in at least one browser caused the
horizontal scroll, when present, to be forced to the end of the line.
Not a side effect I appreciate. Rather than using CSS to alter the
order of presentation, I've decided to alter the order in the HTML.

This change will require that I include an option, when I get there, to
determine if the trigger is placed pre- or post- menu.

As development continues, to be certain that the ID-based selectors are
not being used in styling elements, the menu name will have "x" added
to the one "fixed". The animations, which may validly so selected, will
have to have the same name change applied in the CSS. Removal of all
such name changes will have to be done once I'm sure they're not a
factor in the styling of elements.

I've also noticed that the "handle" for the side menus, when in motion,
floats away from the border line of the menu it controls. Once fully
open it is attached, as it should be. Something to address while
working on the styles for the menus and trigger.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>