Adjust reactiveness for burger icons
Adjust the relative measure usage of the burger icon so that it still
has a good presentation at very high zoom factors. Up to a font size of
72, the Firefox max, or 450% zoom, the icon grows, and stays reasonably
operational. Though, at font sizes below 13 the "X" of an open menu is
slightly out of balance. The likelyhood of such a use case on a low res
device is rare at best, if even possible.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>